feat: Reorganize library in preparation for ini filtering
